category Repo

+++ category-service/src/main/java/com/rancho/service/impl/CategoryServiceImpl.java
@@ -42,9 +42,11 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategoryById(Long id) throws Exception {
-
-        getCategoryById(id);
+    public void deleteCategoryById(Long id, Long salonId) throws Exception {
+        Category category=getCategoryById(id);
+        if(category.getSalonId().equals(salonId)){
+            throw new Exception("you don't have permission to delete this category")
+        }
         categoryRepository.deleteById(id);
